Version 2.0.0
Currently we can only have one key pair for every endpoint we want to do
request signing and verification. This is an issue as we need to share
the same signing key to all our clients.

We need a new way to allow a verification server to be able to verify
different signing keys.

In Version 2.0.0 we have have introduced a concept of a `KeyStore` where
you can add signing and verification keys to.

For example when signing requests:

```ruby
private_key = <<-pem.gsub(/^\s+/, "")
    -----BEGIN EC PRIVATE KEY-----
    MHcCAQEEIBOQ3YIILYMV1glTKbF9oeZWzHe3SNQjAx4IbPIxNygQoAoGCCqGSM49
    AwEHoUQDQgAEuOC3ufTTnW0hVmCPNERb4LxaDE/OexDdlmXEjHYaixzYIduluGXd
    3cjg4H2gjqsY/NCpJ9nM8/AAINSrq+qPuA==
    -----END EC PRIVATE KEY-----
  pem

require 'openssl'

JWTSignedRequest.configure_keys do |config|
  config.add_signing_key(
    key_id: 'client_a',
    key: OpenSSL::PKey::EC.new(private_key),
    algorithm: 'ES256',
  )
end

jwt_token = JWTSignedRequest.sign(
  method: 'GET',
  path: '/api',
  headers: {"Content-Type" => "application/json"},
  body: "",
  key_id: 'client_a',
)
```

Now when signing requests we only have to specify the key_id we are
using.

And on when verifying requests:

```ruby
public_key = <<-pem.gsub(/^\s+/, "")
  -----BEGIN PUBLIC KEY-----
  MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEuOC3ufTTnW0hVmCPNERb4LxaDE/O
  exDdlmXEjHYaixzYIduluGXd3cjg4H2gjqsY/NCpJ9nM8/AAINSrq+qPuA==
  -----END PUBLIC KEY-----
pem

require 'openssl'

JWTSignedRequest.configure_keys do |config|
  config.add_verification_key(
    key_id: 'client_a',
    key: OpenSSL::PKey::EC.new(public_key),
    algorithm: 'ES256',
  )
end

JWTSignedRequest.verify(request: request)
```

Now when signing requests we are getting the key_id that was used to sign the JWT
token. This key_id is used to find the corresponding verification key
from the key store.