Fix changelog order
From: https://keepachangelog.com/en/1.0.0/

> * The latest version comes first.

I read a lot of changelogs every week and this one is upside-down
compared to the convention I see most commonly.

I'd appreciate if we flip the order so the the stuff people care about
is in front of them as soon as the changelog loads/opens.