Add support for JWT version 2.1
Newer versions of JWT require that the signing algorithm must be specified
when verifying a JWT. https://github.com/jwt/ruby-jwt/commit/0a2b94433c4626dae797de68d0f530c3cf6fed9d

When verifying the request we now use the algorithm of the
stored key.

An exception will be raised when using JWT 2.x when attempting to use a
stored key that does not specify an algorithm.

The deprecated argument `algorithm` for the `JWTSignedRequest::Verify.initialize` method was previously redundant and setting it did nothing. This argument if set would effectively override the algorithm for the found stored_key. This however is the same behaviour of the other deprecated argument `secret_key` and seems somewhat consistent.