Add compatability for 'faraday' >= 1.2.0.
The commit
https://github.com/lostisland/faraday/commit/d56742a174fe9f2f9f3d46438324fcac0d2381f7,
which was introduced in version 1.2.0, broke the current Faraday middleware.

This is because the implementation of the Faraday::Middleware class changed in
such a way where @options was being set when #initialize was called.

In the JWTSignedRequest::Middlewares::Faraday class we set @options, and this
was being overridden by Faraday::Middleware#initialize.

With this change we check the version of Faraday we are using and pass the
options hash in to the call to super if it is a version that requires it.

I have updated the Appraisals file to check compatibility with the last 4 minor
versions of Faraday.